// src/store/actions/product/product-action-type.ts
export enum ProductActionType {
    CREATE_PRODUCT_REQUEST = 'CREATE_PRODUCT_REQUEST',
    CREATE_PRODUCT_SUCCESS = 'CREATE_PRODUCT_SUCCESS',
    CREATE_PRODUCT_FAILURE = 'CREATE_PRODUCT_FAILURE',

    GET_PRODUCTS_REQUEST = 'GET_PRODUCTS_REQUEST',
    GET_PRODUCTS_SUCCESS = 'GET_PRODUCTS_SUCCESS',
    GET_PRODUCTS_FAILURE = 'GET_PRODUCTS_FAILURE',

    REMOVE_PRODUCT_REQUEST = 'REMOVE_PRODUCT_REQUEST',
    REMOVE_PRODUCT_FAILURE = 'REMOVE_PRODUCT_FAILURE',
    REMOVE_PRODUCT_SUCCESS = 'REMOVE_PRODUCT_SUCCESS',

    UPDATE_PRODUCT_REQUEST = 'UPDATE_PRODUCT_REQUEST',
    UPDATE_PRODUCT_FAILURE = 'UPDATE_PRODUCT_FAILURE',
    UPDATE_PRODUCT_SUCCESS = 'UPDATE_PRODUCT_SUCCESS',
}